backend/requirements.txt
Flask
flask-cors
pymongo
bcrypt
pyjwt
cloudinary
flask-jwt-extended
google-generativeai 
sentence-transformers
python-dotenv
python-dateutil
pandas
numpy
openpyxl
scikit-learn
requests
python-docx
PyPDF2
pdfplumber
torch
transformers
huggingface-hub
safetensors
tokenizers
scipy
Pillow
filelock
packaging
pyyaml
regex
tqdm
protobuf
google-api-core
google-api-python-client
google-auth
google-ai-generativelanguage
proto-plus
googleapis-common-protos
grpcio
grpcio-status
cachetools
requests-oauthlib
authlib
pyasn1-modules
pyasn1
rsa
pytz
tzdata
lxml
pdfminer.six
pypdfium2
cryptography
cffi
pycparser
colorama
sympy
networkx
mpmath
httplib2
google-auth-httplib2
uritemplate
pyparsing
annotated-types
pydantic-core
typing-inspection
dnspython
six
jinja2
itsdangerous
click
blinker
flask-socketio
redis
gunicorn